+++ src/ssh-tunneling.test.ts
@@ -1,3 +1,5 @@
+import assert from 'assert'
+import {ChildProcess} from 'child_process'
 import {Server, Socket} from 'net'
 import {Client as SshClient, ClientChannel} from 'ssh2'
 import {
@@ -214,7 +216,8 @@ describe('openSshTunnel', () => {
     const [event, listener] = capture(mockSshClientType.on).last()
     expect(event).to.equal('ready')
 
-    listener()
+    const sshClientListener = (listener as unknown) as (() => void)
+    sshClientListener()
 
     verify(mockReadyListenerContainerType.func(result)).once()
   })
@@ -234,6 +237,7 @@ describe('openSshTunnel', () => {
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const [_, _1, _2, _3, portForwardListener] = capture(mockSshClientType.forwardOut).last()
+    assert(typeof portForwardListener !== 'undefined')
     portForwardListener(undefined, mockSshStreamInstance)
 
     verify(mockTcpSocketType.pipe(mockSshStreamInstance)).once()
@@ -259,6 +263,7 @@ describe('openSshTunnel', () => {
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const [_, _1, _2, _3, portForwardListener] = capture(mockSshClientType.forwardOut).last()
+    assert(typeof portForwardListener !== 'undefined')
     portForwardListener(undefined, mockSshStreamInstance)
 
     verify(mockTcpSocketType.pipe(mockSshStreamInstance)).once()
@@ -338,6 +343,7 @@ describe('openSshTunnel', () => {
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const [_, _1, _2, _3, portForwardListener] = capture(mockSshClientType.forwardOut).last()
+    assert(typeof portForwardListener !== 'undefined')
 
     const fakeError = new Error('Just testing!')
     try {
@@ -424,7 +430,7 @@ describe('openSshTunnel', () => {
 
 describe('tunnel services', () => {
   it('should have a valid child process factory', () => {
-    let result = null
+    let result: ChildProcess | null = null
     try {
       result = tunnelServices.childProcessFactory.spawn('ls', {})